+++ survival-graphics/src/main/java/net/survival/graphics/TextStyle.kt
@@ -1,25 +1,25 @@
 package net.survival.graphics
 
-private const val DEFAULT_RED = 1.0
-private const val DEFAULT_GREEN = 1.0
-private const val DEFAULT_BLUE = 1.0
-private const val DEFAULT_ALPHA = 1.0
-private const val DEFAULT_FONT_SIZE = 1.0
-private const val DEFAULT_HORIZONTAL_SPACING = 0.125
-private const val DEFAULT_VERTICAL_SPACING = 0.25
-private const val DEFAULT_SPACE_WIDTH = 0.5
-private const val DEFAULT_TAB_WIDTH = 2.0
+private const val DEFAULT_RED = 1.0f
+private const val DEFAULT_GREEN = 1.0f
+private const val DEFAULT_BLUE = 1.0f
+private const val DEFAULT_ALPHA = 1.0f
+private const val DEFAULT_FONT_SIZE = 1.0f
+private const val DEFAULT_HORIZONTAL_SPACING = 0.125f
+private const val DEFAULT_VERTICAL_SPACING = 0.25f
+private const val DEFAULT_SPACE_WIDTH = 0.5f
+private const val DEFAULT_TAB_WIDTH = 2.0f
 
 class TextStyle private constructor(
-    val red: Double,
-    val green: Double,
-    val blue: Double,
-    val alpha: Double,
-    val fontSize: Double,
-    val horizontalSpacing: Double,
-    val verticalSpacing: Double,
-    val spaceWidth: Double,
-    val tabWidth: Double
+    val red: Float,
+    val green: Float,
+    val blue: Float,
+    val alpha: Float,
+    val fontSize: Float,
+    val horizontalSpacing: Float,
+    val verticalSpacing: Float,
+    val spaceWidth: Float,
+    val tabWidth: Float
 ) {
     companion object {
         val DEFAULT = TextStyle(
